requirements.txt
fastapi
uvicorn[standard]
python-docx
pyyaml
requests
PyJWT
python-multipart
pydantic[email]
reportlab
